refactor S3Restore out of S3BackingStore.scala
